LayoutTests/fast/events/spatial-navigation/snav-iframe-no-focusable-content-expected.txt

a


e
PASS document.activeElement.getAttribute("id") is "end"
PASS document.activeElement.getAttribute("id") is "start"
